fix: major version bump compatibility (rusqlite 0.40, keyring 4.x, deps)
- Workspace conflict: rusqlite 0.31 (services) vs 0.40 (core)
  unified to 0.40 across all packages; deleted stale lockfiles and
  regenerated from scratch
- telemetry db.rs: u64 count not implementing ToSql, cast to i64
- keychain.rs: keyring 4.x renamed delete_password() to
  delete_credential()
- zustand 5.x: react 19 peer dep added
- Cargo.lock regenerated across all packages

"

+++ services/telemetry/src/db.rs
@@ -64,7 +64,7 @@ impl Db {
         for ev in events {
             conn.execute(
                 "INSERT INTO usage_events (app_version, os, feature, count, window_ts) VALUES (?1, ?2, ?3, ?4, ?5)",
-                params![ev.app_version, ev.os, ev.feature, ev.count, ev.window_ts],
+                params![ev.app_version, ev.os, ev.feature, ev.count as i64, ev.window_ts],
             )?;
         }
         Ok(())
